Add command line interface for the controller
